requirements.txt
ffmpeg-python==0.2.0
uWSGI==2.0.25.1
python-dotenv==1.0.0